requirements.txt
wget
requests
TgCrypto
youtube_dl
python-dotenv
ffmpeg-python
youtube_search_python
pytgcalls==3.0.0.dev10
git+https://github.com/pyrogram/pyrogram@master
